feat: connect welcome animations with onboarding flow
